refactor(policy): split model family rules into separate files
Spec 1.3 complete: Split defaults/gpt.ts into model family-specific files.

Create policy/defaults/rules/ directory with:
- plans.ts: PAID_PLAN_WEIGHTS, PAID_PLANS, GPT_FREE_BLOCKED_MODELS,
  buildFreeBlockedConstraints
- glm.ts: GLM_PROVIDER_ORDER, createGlmRoutingRules
- claude.ts: CLAUDE_OPUS_46_PROVIDER_ORDER, createClaudeRoutingRules
- gpt.ts: DEFAULT_GPT_PROVIDER_ORDER, GPT_OSS_PROVIDER_ORDER,
  createGptRoutingRules
- index.ts: createAllModelRoutingRules (aggregates in correct order:
  GLM → Claude → GPT)

Update defaults/index.ts to use new structure with backward-compatible
exports. Delete original gpt.ts.

Rule ordering preserved: GLM rules first (most specific), then Claude,
then GPT (catch-all). All 579 tests pass.
